requirements.txt
fastapi[all]==0.116.1
python-jose[cryptography]==3.5.0
requests==2.32.4